feat(twins): boot slack via defineTwin() — sdk-boot contract proof green (FDRS-681)
