+++ knip.json
@@ -27,7 +27,12 @@
   ],
   "workspaces": {
     ".": {
-      "entry": ["scripts/*.{mjs,sh}", "contract/run.mjs", "contract/contract.test.mjs"],
+      "entry": [
+        "scripts/*.{mjs,sh}",
+        "contract/run.mjs",
+        "contract/*.test.mjs",
+        "contract/proof/*.mjs"
+      ],
       "project": ["scripts/**/*.{mjs,sh}", "contract/**/*.mjs"]
     },
     "packages/*": {
